gateway/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
PyJWT==2.9.0
cryptography>=45.0.7,<47
python-dotenv==1.0.1
